refactor:상태를 관리해서 결과를 반환하도록 변경

+++ src/main/kotlin/blackjack/domain/model/Money.kt
@@ -1,7 +1,7 @@
 package blackjack.domain.model
 
 @JvmInline
-value class Money(val value: Int) {
+value class Money(val value: Double) {
     init {
         require(value > 0)
     }
